# custom/mnzl/receivables/openapi/receivables-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MNZL purchased receivables",
    "version": "1"
  },
  "servers": [
    {
      "url": "/fineract-provider/api/v1/mnzl/receivables"
    }
  ],
  "security": [
    {
      "fineractAuthentication": []
    }
  ],
  "paths": {
...
    "/hel-funding/allocations": {
      "post": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/helAdvanceAllocationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helAdvanceAllocationOperation"
                }
              }
            }
          },
          "202": {
            "description": "Accepted replayable operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helAdvanceAllocationOperation"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Operation readback URL"
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
    "/hel-funding/commands": {
      "post": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/helFundingCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helFundingResult"
                }
              }
            }
          },
          "202": {
            "description": "Accepted replayable operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helFundingResult"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Operation readback URL"
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
    "/configuration": {
      "post": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/nativeConfiguration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/capabilities"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
    "/authorizations": {
      "post": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/nativeHistoricalAuthorization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/nativeHistoricalAuthorization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
    "/calculate": {
      "post": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/calculationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/calculationResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
    "/commands": {
      "post": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/financialCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/financialOperation"
                }
              }
            }
          },
          "202": {
            "description": "Accepted replayable operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/financialOperation"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Operation readback URL"
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "fineractAuthentication": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "helAdvanceAllocationCommand": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/helAdvanceAllocationCommand",
        "type": "object",
        "properties": {
          "executionMode": {
            "type": "string",
            "enum": [
              "CURRENT",
              "RECONSTRUCTION",
              "CORRECTION"
            ]
          },
          "executionAuthorization": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "authorizationId": {
                    "$ref": "#/components/schemas/id"
                  },
                  "scopeHash": {
                    "$ref": "#/components/schemas/sha256"
                  },
                  "approvedBy": {
                    "$ref": "#/components/schemas/id"
                  },
                  "effectiveFrom": {
                    "$ref": "#/components/schemas/financialDate"
                  },
                  "effectiveThrough": {
                    "$ref": "#/components/schemas/financialDate"
                  }
                },
                "required": [
                  "authorizationId",
                  "scopeHash",
                  "approvedBy",
                  "effectiveFrom",
                  "effectiveThrough"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "accountMappingRevisionId": {
            "$ref": "#/components/schemas/id"
          },
          "operationId": {
            "$ref": "#/components/schemas/id"
          },
          "idempotencyKey": {
            "$ref": "#/components/schemas/id"
          },
          "expectedVersion": {
            "$ref": "#/components/schemas/version"
          },
          "businessDate": {
            "$ref": "#/components/schemas/financialDate"
          },
          "basisHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "scope": {
            "$ref": "#/components/schemas/scope"
          },
          "subjectId": {
            "$ref": "#/components/schemas/id"
          },
          "subjectKind": {
            "type": "string",
            "const": "HEL_LOAN"
          },
          "affectedAccountVersions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/accountVersion"
            }
          },
          "executionScopeHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "actorId": {
            "$ref": "#/components/schemas/id"
          },
          "approverIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "approvalEvidenceIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "commandType": {
            "type": "string",
            "const": "ALLOCATE_HEL_DEVELOPER_ADVANCE"
          },
          "helFundingOperationId": {
            "$ref": "#/components/schemas/id"
          },
          "dealId": {
            "$ref": "#/components/schemas/id"
          },
          "verifiedAdvanceAllocationIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "payoffMinor": {
            "$ref": "#/components/schemas/positiveMinorUnits"
          },
          "acceptedCustomerTermsHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "settlementApprovalId": {
            "$ref": "#/components/schemas/id"
          }
        },
        "required": [
          "executionMode",
          "executionAuthorization",
          "accountMappingRevisionId",
          "operationId",
          "idempotencyKey",
          "expectedVersion",
          "businessDate",
          "basisHash",
          "scope",
          "subjectId",
          "subjectKind",
          "affectedAccountVersions",
          "executionScopeHash",
          "actorId",
          "approverIds",
          "approvalEvidenceIds",
          "commandType",
          "helFundingOperationId",
          "dealId",
          "verifiedAdvanceAllocationIds",
          "payoffMinor",
          "acceptedCustomerTermsHash",
          "settlementApprovalId"
        ],
        "additionalProperties": false
      },
      "helAdvanceAllocationOperation": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/helAdvanceAllocationOperation",
        "type": "object",
        "properties": {
          "operationId": {
            "$ref": "#/components/schemas/id"
          },
          "commandType": {
            "type": "string",
            "const": "ALLOCATE_HEL_DEVELOPER_ADVANCE"
          },
          "subjectKind": {
            "type": "string",
            "const": "HEL_LOAN"
          },
          "subjectId": {
            "$ref": "#/components/schemas/id"
          },
          "scope": {
            "$ref": "#/components/schemas/financialScope"
          },
          "businessDate": {
            "$ref": "#/components/schemas/financialDate"
          },
          "payloadHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "expectedVersion": {
            "$ref": "#/components/schemas/version"
          },
          "affectedAccountVersions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/accountVersion"
            }
          },
          "executionScopeHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "actorId": {
            "$ref": "#/components/schemas/id"
          },
          "approverIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "state": {
            "$ref": "#/components/schemas/operationState"
          },
          "nativeTransactionIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "financialEventIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "eventWatermark": {
            "$ref": "#/components/schemas/version"
          },
          "reconciliationId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/id"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "code": {
                    "$ref": "#/components/schemas/blockerCode"
                  },
                  "retryClassification": {
                    "type": "string",
                    "enum": [
                      "RETRY_SAME_OPERATION",
                      "READ_BEFORE_RETRY",
                      "CORRECTIVE_COMMAND_REQUIRED",
                      "NEW_APPROVAL_REQUIRED",
                      "NOT_RETRYABLE"
                    ]
                  }
                },
                "required": [
                  "code",
                  "retryClassification"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "$ref": "#/components/schemas/utcInstant"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/utcInstant"
          }
        },
        "required": [
          "operationId",
          "commandType",
          "subjectKind",
          "subjectId",
          "scope",
          "businessDate",
          "payloadHash",
          "expectedVersion",
          "affectedAccountVersions",
          "executionScopeHash",
          "actorId",
          "approverIds",
          "state",
          "nativeTransactionIds",
          "financialEventIds",
          "reconciliationId",
          "error",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "helFundingCommand": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/helFundingCommand",
        "type": "object",
        "properties": {
          "executionMode": {
            "type": "string",
            "enum": [
              "CURRENT",
              "RECONSTRUCTION",
              "CORRECTION"
            ]
          },
          "executionAuthorization": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "authorizationId": {
                    "$ref": "#/components/schemas/id"
                  },
                  "scopeHash": {
                    "$ref": "#/components/schemas/sha256"
                  },
                  "approvedBy": {
                    "$ref": "#/components/schemas/id"
                  },
                  "effectiveFrom": {
                    "$ref": "#/components/schemas/financialDate"
                  },
                  "effectiveThrough": {
                    "$ref": "#/components/schemas/financialDate"
                  }
                },
                "required": [
                  "authorizationId",
                  "scopeHash",
                  "approvedBy",
                  "effectiveFrom",
                  "effectiveThrough"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "accountMappingRevisionId": {
            "$ref": "#/components/schemas/id"
          },
          "operationId": {
            "$ref": "#/components/schemas/id"
          },
          "idempotencyKey": {
            "$ref": "#/components/schemas/id"
          },
          "expectedVersion": {
            "$ref": "#/components/schemas/version"
          },
          "businessDate": {
            "$ref": "#/components/schemas/financialDate"
          },
          "basisHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "scope": {
            "$ref": "#/components/schemas/scope"
          },
          "subjectId": {
            "$ref": "#/components/schemas/id"
          },
          "subjectKind": {
            "type": "string",
            "const": "HEL_LOAN"
          },
          "affectedAccountVersions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/accountVersion"
            }
          },
          "executionScopeHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "actorId": {
            "$ref": "#/components/schemas/id"
          },
          "approverIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "approvalEvidenceIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "commandType": {
            "type": "string",
            "const": "FUND_HEL_TO_SETTLEMENT_CLEARING"
          },
          "applicationId": {
            "$ref": "#/components/schemas/id"
          },
          "dealId": {
            "$ref": "#/components/schemas/id"
          },
          "expectedNativeClientId": {
            "$ref": "#/components/schemas/id"
          },
          "loanExternalId": {
            "$ref": "#/components/schemas/id"
          },
          "expectedPrincipalMinor": {
            "$ref": "#/components/schemas/positiveMinorUnits"
          },
          "expectedFinancedFeesMinor": {
            "$ref": "#/components/schemas/version"
          },
          "financedFeeIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          }
        },
        "required": [
          "executionMode",
          "executionAuthorization",
          "accountMappingRevisionId",
          "operationId",
          "idempotencyKey",
          "expectedVersion",
          "businessDate",
          "basisHash",
          "scope",
          "subjectId",
          "subjectKind",
          "affectedAccountVersions",
          "executionScopeHash",
          "actorId",
          "approverIds",
          "approvalEvidenceIds",
          "commandType",
          "applicationId",
          "dealId",
          "expectedNativeClientId",
          "loanExternalId",
          "expectedPrincipalMinor",
          "expectedFinancedFeesMinor",
          "financedFeeIds"
        ],
        "additionalProperties": false
      },
...
      "helFundingResult": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/helFundingResult",
        "type": "object",
        "properties": {
          "operation": {
            "$ref": "#/components/schemas/helFundingOperation"
          },
          "nativeLoanId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/id"
              },
              {
                "type": "null"
              }
            ]
          },
          "nativeLoanStatus": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PENDING",
                  "ACTIVE",
                  "CLOSED"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "clearingAmountMinor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/version"
              },
              {
                "type": "null"
              }
            ]
          },
          "nativeTransactionIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          },
          "journalIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/id"
            }
          }
        },
        "required": [
          "operation",
          "nativeLoanId",
          "nativeLoanStatus",
          "clearingAmountMinor",
          "nativeTransactionIds",
          "journalIds"
        ],
        "additionalProperties": false
      },
...
      "nativeHistoricalAuthorization": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/nativeHistoricalAuthorization",
        "type": "object",
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/financierScope"
          },
          "authorizationId": {
            "$ref": "#/components/schemas/id"
          },
          "scopeHash": {
            "$ref": "#/components/schemas/sha256"
          },
          "approvedBy": {
            "$ref": "#/components/schemas/id"
          },
          "effectiveFrom": {
            "$ref": "#/components/schemas/financialDate"
          },
          "effectiveThrough": {
            "$ref": "#/components/schemas/financialDate"
          },
          "mode": {
            "type": "string",
            "enum": [
              "RECONSTRUCTION",
              "CORRECTION"
            ]
          }
        },
        "required": [
          "scope",
          "authorizationId",
          "scopeHash",
          "approvedBy",
          "effectiveFrom",
          "effectiveThrough",
          "mode"
        ],
        "additionalProperties": false
      }
    }
  }
}